Add files via upload
Newest version now plays as a best of 3 rounds, includes a scoreboard to keep track of total matches won, added images to the results window by converting it from a message box popup to a toplevel window, also put an exit button on the new window. added a popup when a winner is determined declaring who won the match.